more testing, moved badvalueutils
