+++ validation/src/test/scala/hmda/validation/rules/lar/validity/V255Spec.scala
@@ -2,7 +2,7 @@ package hmda.validation.rules.lar.validity
 
 import hmda.model.fi.lar.LoanApplicationRegister
 import hmda.validation.rules.EditCheck
-import hmda.validation.rules.lar.LarEditCheckSpec
+import hmda.validation.rules.lar.{BadValueUtils, LarEditCheckSpec}
 
 class V255Spec extends LarEditCheckSpec with BadValueUtils {
   property("Succeeds when Action Taken Type = 1, 2, 3, 4, 5, 6, 7, or 8") {
