+++ validation/src/test/scala/hmda/validation/rules/lar/validity/V355Spec.scala
@@ -1,9 +1,9 @@
 package hmda.validation.rules.lar.validity
 
 import hmda.model.fi.lar.LoanApplicationRegister
-import hmda.validation.dsl.{ Failure, Success }
+import hmda.validation.dsl.{Failure, Success}
 import hmda.validation.rules.EditCheck
-import hmda.validation.rules.lar.LarEditCheckSpec
+import hmda.validation.rules.lar.{BadValueUtils, LarEditCheckSpec}
 import org.scalacheck.Gen
 
 class V355Spec extends LarEditCheckSpec with BadValueUtils {
