+++ validation/src/test/scala/hmda/validation/rules/lar/validity/V425Spec.scala
@@ -1,9 +1,9 @@
 package hmda.validation.rules.lar.validity
 
-import hmda.model.fi.lar.{ Loan, LoanApplicationRegister }
-import hmda.validation.dsl.{ Failure, Success }
+import hmda.model.fi.lar.{Loan, LoanApplicationRegister}
+import hmda.validation.dsl.{Failure, Success}
 import hmda.validation.rules.EditCheck
-import hmda.validation.rules.lar.LarEditCheckSpec
+import hmda.validation.rules.lar.{BadValueUtils, LarEditCheckSpec}
 import org.scalacheck.Gen
 
 class V425Spec extends LarEditCheckSpec with BadValueUtils {
