+++ validation/src/test/scala/hmda/validation/rules/lar/validity/V535Spec.scala
@@ -2,7 +2,7 @@ package hmda.validation.rules.lar.validity
 
 import hmda.model.fi.lar.LoanApplicationRegister
 import hmda.validation.rules.EditCheck
-import hmda.validation.rules.lar.LarEditCheckSpec
+import hmda.validation.rules.lar.{BadValueUtils, LarEditCheckSpec}
 import org.scalacheck.Gen
 
 class V535Spec extends LarEditCheckSpec with BadValueUtils {
